+++ client/src/App.js
@@ -4,7 +4,7 @@ import { getMessage } from "./service";
 import { BrowserRouter, Route, Link, Switch } from "react-router-dom";
 import SignupForm from "./components/SignupForm";
 import LoginForm from "./components/LoginForm";
-import Skills from "./Pages/Skills";
+import SkillsNav from "./Pages/SkillsNav";
 import Modal from "./components/Modal";
 import MentorsView from "./Pages/MentorsView";
 
@@ -26,7 +26,7 @@ export function App() {
             <Route exact path="/" component={LoginForm} />
             <Route path="/signup" component={SignupForm} />
             {/* <LoginForm /> */}
-            <Route path="/skills" component={Skills} />
+            <Route path="/skills" component={SkillsNav} />
             <Route path="/modal" component={Modal} />
             <Route path="/MentorsView" component={MentorsView} />
           </div>
