+++ server/db.js
@@ -14,7 +14,7 @@ if (process.env.DATABASE_URL) { //it's set in Heroku
   }
 } else { //default local config
   config = {
-  user: 'Altom',
+  user: '',
   host: 'localhost',
   database: 'cyf_knowledge_checklist',
   password: '123',
